+++ src/konpy/views.clj
@@ -2,7 +2,7 @@
   (:require [hiccup2.core :as h]
             [ring.util.response :as response]))
 
-(def ^:private version "0.7.8")
+(def ^:private version "0.7.9-SNAPSHOT")
 
 (defn base
   [content]